+++ src/main/scala/de/unruh/isabelle/pure/Term.scala
@@ -2,10 +2,10 @@ package de.unruh.isabelle.pure
 
 import de.unruh.isabelle.control.Isabelle.{DInt, DList, DObject, DString}
 import de.unruh.isabelle.control.{Isabelle, OperationCollection}
-import de.unruh.isabelle.misc.{Symbols, Utils}
+import de.unruh.isabelle.misc.{FutureValue, Symbols, Utils}
 import de.unruh.isabelle.mlvalue.MLValue.Converter
 import de.unruh.isabelle.mlvalue.Implicits._
-import de.unruh.isabelle.mlvalue.{FutureValue, MLFunction, MLFunction2, MLFunction3, MLRetrieveFunction, MLValue}
+import de.unruh.isabelle.mlvalue.{MLFunction, MLFunction2, MLFunction3, MLRetrieveFunction, MLValue}
 import de.unruh.isabelle.pure.Implicits._
 import de.unruh.isabelle.pure.Term.Ops
 import org.apache.commons.lang3.builder.HashCodeBuilder